requirements.txt
# RAG System Dependencies
# FDA 510(k) Feedback Response System

# Core
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0

# Vector Store
chromadb>=0.4.22

# Document Loaders
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.2
unstructured>=0.12.0

# Embeddings
openai>=1.12.0
tiktoken>=0.5.2

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
rich>=13.7.0

# For hybrid search
rank-bm25>=0.2.2

# CLI
click>=8.1.7
